Set date/time fields by numeric value
This makes handling of date and time fields symmetric, since we've always
fetched it numerically; it also offloads a small amout of processing from the
framework and into the database.

Additionally this fixes undiscovered bugs that would have cropped up if mysql
and php disagreed about the timezone.

Furthermore, this moves the "what type of field is it" out of the codepath that
gets hit everytime an object is updated or created and into the cached per-class
codepath.