Remove concept of current statement and statement stack in Database
Just use a PDOStatement handle the way it's intended, don't overcomplicate things.

PDOStatement provides far more robust methods than what was implemented prior to
this round of ripping things out, jsut RTFM PDOStatement and use it instead of
working around it.

This leads to easier to write and understand codepaths in fewer lines since
there's no need to free some redundant stashed handle.

Not only does this amke Database easier to understand, but it makes it far more
flexible since it's now "just" a connection manager with a smattering of
convenience methods for creating handles.